Split the written code into separate modules
